Delete operation bug fixed

+++ controllers/commentController.js
@@ -88,11 +88,20 @@ const deleteComment = async (req, res, next) => {
     const commentId = req.params.commentId;
     const db = firestore.getFirestore(firebase);
     const comment = await firestore.doc(db, "comments", commentId);
-    await firestore.deleteDoc(comment);
+    const commentCheck = await firestore.getDoc(comment);
 
-    res.status(200).json({
-      message: "Comment record has been deleted successfully!",
-    });
+    if (commentCheck.exists()) {
+      await firestore.deleteDoc(comment);
+      res.status(200).json({
+        message: "Comment record has been deleted successfully!",
+      });
+    } else {
+      let errorMessage = "Comment cannot found.";
+      logger.error(errorMessage);
+      res.status(500).json({
+        message: errorMessage,
+      });
+    }
   } catch (error) {
     logger.error(error.message);
     res.status(500).json({
