+++ controllers/newController.js
@@ -89,11 +89,20 @@ const deleteNew = async (req, res, next) => {
         const newId = req.params.newId;
         const db = firestore.getFirestore(firebase);
         const _new = await firestore.doc(db, "news", newId);
-        await firestore.deleteDoc(_new);
+        const newCheck = await firestore.getDoc(_new);
 
-        res.status(200).json({
-            message: "New record has been deleted successfully!",
-        });
+        if (newCheck.exists()) {
+            await firestore.deleteDoc(_new);
+            res.status(200).json({
+                message: "New record has been deleted successfully!",
+            });
+        } else {
+            let errorMessage = "New cannot found.";
+            logger.error(errorMessage);
+            res.status(500).json({
+                message: errorMessage,
+            });
+        }
     } catch (error) {
         logger.error(error.message);
         res.status(500).json({
