+++ controllers/postController.js
@@ -89,11 +89,20 @@ const deletePost = async (req, res, next) => {
         const postId = req.params.postId;
         const db = firestore.getFirestore(firebase);
         const post = await firestore.doc(db, "posts", postId);
-        await firestore.deleteDoc(post);
+        const postCheck = await firestore.getDoc(post);
 
-        res.status(200).json({
-            message: "Post record has been deleted successfully!",
-        });
+        if (postCheck.exists()) {
+            await firestore.deleteDoc(post);
+            res.status(200).json({
+                message: "Post record has been deleted successfully!",
+            });
+        } else {
+            let errorMessage = "Post cannot found.";
+            logger.error(errorMessage);
+            res.status(500).json({
+                message: errorMessage,
+            });
+        }
     } catch (error) {
         logger.error(error.message);
         res.status(500).json({
